Remove trap & Remove unused function & Separate code into two files
