+++ core/src/executive/vm_exec.rs
